+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # CHANGELOG
 
 ## 0.4.7-SNAPSHOT
-
+* The stats URI for computes available at /compute/stats
+no longer have the raw metrics (CPU utilization, memory) etc
+available at this endpoint. The only in-memory metric available 
+at this URI will be the last successful collection time.
 
 ## 0.4.6